+++ doc/message_target.lua
@@ -1,7 +1,7 @@
 -- message_target
 -- @short: send a utf-8 coded text message to a frameserver
--- @inargs: tgtvid, msg
--- @outargs: bytes_left
+-- @inargs: vid:tgtvid, string:msg
+-- @outargs: int:bytes_left
 -- @longdescr: In some rare occasions (custom applications and clipboard
 -- not using bchunk- method of transfer) it makes sense to send short custom
 -- text messages to a target frameserver. This function takes a string,
@@ -10,6 +10,9 @@
 -- the entire message was sent, a negative value if the string failed to
 -- validate or a positive number indicating the number of bytes left. The
 -- latter case only occurs if the target event queue is full.
+-- @note: In cases where an arcan appl was spawned by another arcan instance,
+-- WORLDID is a valid vid for message_target to the parent. This can be tested
+	-- via ref:valid_vid(WORLDID, TYPE_FRAMESERVER).
 -- @group: targetcontrol
 -- @cfunction: targetmessage
 -- @related: target_input