+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.6.3
+## Lua
+ * nbio-linebuffer callback read truncation edge case fixed
+ * message\_target and valid\_vid now work on WORLDID for LWA
+ * event handler can now be attached to WORLDID for outer WM integration
+ * more shmif events mapped to arcantarget handlers
+
 ## 0.6.2
 
 ## Net